Update
Added viewport scaler to show the player a greater field of view when landed on a pillar. Includes a smooth transition curve.

Added game state function to start/stop and change settings without reloading the page.

Added modal templates for the settings, rocket picker, and graphing utility to see the trajectory.